refactor: update historical contribution fetcher and integrate with main script
