+++ scripts/utils/github-helpers.js
@@ -13,71 +13,6 @@ function getLinkedIssueNumbers(prBody) {
 }
 
 /**
- * HELPER: Discover the user's GitHub join year.
- */
-async function getGitHubJoinYear(axiosInstance, username) {
-  try {
-    const response = await axiosInstance.get(`/users/${username}`);
-    const joinDate = new Date(response.data.created_at);
-    return joinDate.getFullYear();
-  } catch (err) {
-    console.error(
-      `❌ Error discovering GitHub join date for ${username}, defaulting to 2020:`,
-      err.message
-    );
-    return 2020;
-  }
-}
-
-/**
- * HELPER: Fetches the date of the user's first review on a specific PR.
- */
-async function getPrMyFirstReviewDate(owner, repo, prNumber, username, axiosInstance) {
-  try {
-    const response = await axiosInstance.get(`/repos/${owner}/${repo}/pulls/${prNumber}/reviews`);
-    const myReviews = response.data
-      .filter((review) => review.user?.login === username)
-      .sort((a, b) => new Date(a.submitted_at) - new Date(b.submitted_at));
-    if (myReviews.length > 0) {
-      return myReviews[0].submitted_at;
-    }
-    return null;
-  } catch (err) {
-    if (err.response && (err.response.status === 403 || err.response.status === 404)) {
-      return null;
-    }
-    throw err;
-  }
-}
-
-/**
- * HELPER: Fetches the date of the user's first comment on an issue or PR.
- */
-async function getFirstCommentDate(url, username, axiosInstance) {
-  try {
-    let page = 1;
-    while (true) {
-      const response = await axiosInstance.get(`${url}?per_page=100&page=${page}`);
-      const myFirstComment = response.data.find((comment) => comment.user?.login === username);
-      if (myFirstComment) {
-        return myFirstComment.created_at;
-      }
-      const linkHeader = response.headers.link;
-      if (linkHeader && linkHeader.includes('rel="next"')) {
-        page++;
-      } else {
-        return null;
-      }
-    }
-  } catch (err) {
-    if (err.response && (err.response.status === 403 || err.response.status === 404)) {
-      return null;
-    }
-    throw err;
-  }
-}
-
-/**
  * HELPER: Fetches specific activity metadata for a PR to determine "Who has the ball".
  */
 async function getPrActivityMeta(owner, repo, prNumber, axiosInstance, prMainUpdatedAt) {
@@ -161,8 +96,5 @@ async function getPrActivityMeta(owner, repo, prNumber, axiosInstance, prMainUpd
 
 module.exports = {
   getLinkedIssueNumbers,
-  getGitHubJoinYear,
-  getPrMyFirstReviewDate,
-  getFirstCommentDate,
   getPrActivityMeta,
 };
